use ginkgo CLI
Cleanup controller-runtime’s envtest environment to run e2e tests as it is not being used
