+++ .github/workflows/test.yaml
@@ -53,7 +53,6 @@ jobs:
         exclude:
           - istio-type: sail
             pr-event: true
-      fail-fast: false
     runs-on: ubuntu-latest
     env:
       KIND_CLUSTER_NAME: kuadrant-test
